+++ src/Tex2ss/Generator.hs
@@ -1,38 +1,34 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeApplications #-}
 
 module Tex2ss.Generator
-  ( GeneratorResult (..)
+  ( AssembledSource (..)
+  , GeneratedContent (..)
+  , GeneratorResult (..)
   , assembleGeneratedSource
+  , pandocFragmentMarker
   , runPreGenerator
   ) where
 
 import Control.Exception (try)
-import Control.Monad (unless, when)
+import Control.Monad (unless)
 import Control.Monad.Except (throwError)
-import Data.Aeson
-  ( FromJSON (parseJSON)
-  , Object
-  , ToJSON (toJSON)
-  , object
-  , withObject
-  , withText
-  , (.:)
-  , (.=)
-  )
-import qualified Data.Aeson.Key as Key
-import qualified Data.Aeson.KeyMap as KeyMap
-import qualified Data.Aeson.Types as Aeson
-import Data.Foldable (traverse_)
-import qualified Data.Map.Strict as Map
+import Data.Aeson (ToJSON (toJSON), object, (.=))
+import qualified Data.ByteString.Char8 as ByteString
+import Data.Char (ord)
 import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import Data.Text (Text)
 import qualified Data.Text as Text
 import qualified HsLua as Lua
+import Numeric (showHex)
 import Text.Pandoc (PandocError, runIO)
+import Text.Pandoc.Definition (Block (RawBlock), Format (..), Inline (RawInline))
 import Text.Pandoc.Error (renderError)
 import Text.Pandoc.Lua (Global (PANDOC_SCRIPT_FILE), runLua, setGlobals)
+import Text.Pandoc.Walk (query)
 import Tex2ss.Diagnostics (Diagnostic, Severity (Error), diagnosticAt)
 import Tex2ss.Types
   ( Bundle (bundleIndexPath, bundleSlot)
@@ -41,27 +37,21 @@ import Tex2ss.Types
   , renderSlot
   )
 
+data GeneratedContent
+  = DeferredLaTeXBlock Text
+  | PandocBlocks [Block]
+  deriving stock (Eq, Show)
+
 newtype GeneratorResult = GeneratorResult
-  { generatedFragments :: Map Text Text
+  { generatedFragments :: Map Text GeneratedContent
   }
   deriving stock (Eq, Show)
 
-data DeferredFragment = DeferredFragment Text
-
-instance FromJSON DeferredFragment where
-  parseJSON = withObject "generated fragment" $ \value -> do
-    rejectUnknown "generated fragment" (Set.fromList ["type", "value"]) value
-    kind <- value .: "type" >>= withText "fragment type" pure
-    when (kind /= "deferred_latex") $
-      fail "fragment type must be deferred_latex in this experiment"
-    DeferredFragment <$> value .: "value"
-
-instance FromJSON GeneratorResult where
-  parseJSON = withObject "generator result" $ \value -> do
-    rejectUnknown "generator result" (Set.singleton "fragments") value
-    fragments <- value .: "fragments"
-    traverse_ validateFragmentName (Map.keys fragments)
-    pure . GeneratorResult $ fmap (\(DeferredFragment body) -> body) fragments
+data AssembledSource = AssembledSource
+  { assembledText :: Text
+  , assembledPandocFragments :: Map Text [Block]
+  }
+  deriving stock (Eq, Show)
 
 data GeneratorContext = GeneratorContext
   { contextCurrentSlot :: Text
@@ -92,7 +82,7 @@ runPreGenerator scriptPath siteIndex bundle = do
         Lua.TypeFunction -> do
           Lua.pushViaJSON context
           Lua.callTrace 1 1
-          Lua.forcePeek $ Lua.peekViaJSON Lua.top `Lua.lastly` Lua.pop 1
+          Lua.forcePeek $ peekGeneratorResult Lua.top `Lua.lastly` Lua.pop 1
         Lua.TypeNil -> do
           Lua.pop 1
           Lua.failLua "generator must define pre_generator(context)"
@@ -106,16 +96,104 @@ runPreGenerator scriptPath siteIndex bundle = do
       Right result ->
         case result of
           Left problem -> Left [generatorDiagnostic bundle scriptPath problem]
-          Right generated -> Right generated
+          Right generated -> validateGeneratorResult scriptPath generated
 
-assembleGeneratedSource :: FilePath -> GeneratorResult -> Text -> Either [Diagnostic] Text
-assembleGeneratedSource sourcePath result source =
-  fmap (Text.intercalate "\n") . traverse replaceLine $ Text.splitOn "\n" source
+peekGeneratorResult :: Lua.LuaError error => Lua.Peeker error GeneratorResult
+peekGeneratorResult index = do
+  rejectUnknownFields "generator result" (Set.singleton "fragments") index
+  GeneratorResult
+    <$> Lua.peekFieldRaw
+      (Lua.peekMap Lua.peekText peekGeneratedContent)
+      "fragments"
+      index
+
+peekGeneratedContent :: Lua.LuaError error => Lua.Peeker error GeneratedContent
+peekGeneratedContent index = do
+  kind <- Lua.peekFieldRaw Lua.peekText "type" index
+  case kind of
+    "deferred_latex" -> do
+      rejectUnknownFields "deferred_latex fragment" (Set.fromList ["type", "value"]) index
+      DeferredLaTeXBlock <$> Lua.peekFieldRaw Lua.peekText "value" index
+    "pandoc_blocks" -> do
+      rejectUnknownFields "pandoc_blocks fragment" (Set.fromList ["type", "blocks"]) index
+      PandocBlocks
+        <$> Lua.peekFieldRaw
+          (Lua.peekList (Lua.safepeek @Block))
+          "blocks"
+          index
+    _ ->
+      Lua.failPeek . ByteString.pack $
+        "fragment type must be deferred_latex or pandoc_blocks"
+
+rejectUnknownFields
+  :: Lua.LuaError error
+  => String
+  -> Set.Set Text
+  -> Lua.Peeker error ()
+rejectUnknownFields label allowed index = do
+  present <-
+    Map.keysSet
+      <$> Lua.peekMap Lua.peekText (const $ pure ()) index
+  let unknown = Set.toAscList (present `Set.difference` allowed)
+  unless (null unknown) . Lua.failPeek . ByteString.pack $
+    label <> " contains unknown fields: " <> Text.unpack (Text.intercalate ", " unknown)
+
+validateGeneratorResult :: FilePath -> GeneratorResult -> Either [Diagnostic] GeneratorResult
+validateGeneratorResult scriptPath result@(GeneratorResult fragments) =
+  case nameProblems <> rawProblems of
+    [] -> Right result
+    problems -> Left problems
+ where
+  nameProblems =
+    [ diagnosticAt
+        Error
+        "generator.fragment-name-invalid"
+        scriptPath
+        ("fragment name must match [a-z0-9][a-z0-9_-]*: " <> name)
+    | name <- Map.keys fragments
+    , not (validFragmentName name)
+    ]
+  rawProblems =
+    [ diagnosticAt
+        Error
+        "generator.pandoc-blocks-raw"
+        scriptPath
+        ( "pandoc_blocks fragment "
+            <> name
+            <> " contains target-specific raw content: "
+            <> kind
+            <> "["
+            <> format
+            <> "] "
+            <> Text.take 100 (Text.unwords $ Text.words body)
+        )
+    | (name, PandocBlocks blocks) <- Map.toAscList fragments
+    , (kind, format, body) <- generatedRawNodes blocks
+    ]
+
+generatedRawNodes :: [Block] -> [(Text, Text, Text)]
+generatedRawNodes blocks = query blockRaw blocks <> query inlineRaw blocks
+ where
+  blockRaw = \case
+    RawBlock (Format format) body -> [("block", format, body)]
+    _ -> []
+  inlineRaw = \case
+    RawInline (Format format) body -> [("inline", format, body)]
+    _ -> []
+
+assembleGeneratedSource :: FilePath -> GeneratorResult -> Text -> Either [Diagnostic] AssembledSource
+assembleGeneratedSource sourcePath result source = do
+  replaced <- traverse replaceLine $ Text.splitOn "\n" source
+  pure
+    AssembledSource
+      { assembledText = Text.intercalate "\n" (map fst replaced)
+      , assembledPandocFragments = Map.unions (map snd replaced)
+      }
  where
   fragments = generatedFragments result
   replaceLine line =
     case parsePlaceholder line of
-      NoPlaceholder -> Right line
+      NoPlaceholder -> Right (line, Map.empty)
       InvalidPlaceholder message ->
         Left [diagnosticAt Error "generator.placeholder-invalid" sourcePath message]
       Placeholder name ->
@@ -128,7 +206,15 @@ assembleGeneratedSource sourcePath result source =
                   sourcePath
                   ("generator did not return fragment: " <> name)
               ]
-          Just body -> Right body
+          Just (DeferredLaTeXBlock body) -> Right (body, Map.empty)
+          Just (PandocBlocks blocks) ->
+            Right ("\n" <> pandocFragmentMarker name <> "\n", Map.singleton name blocks)
+
+pandocFragmentMarker :: Text -> Text
+pandocFragmentMarker name =
+  "texssgeneratedpandocblocks" <> Text.concatMap encodeCharacter name
+ where
+  encodeCharacter = Text.pack . (<> "z") . (`showHex` "") . ord
 
 data Placeholder
   = NoPlaceholder
@@ -153,11 +239,6 @@ parsePlaceholder line
   stripped = Text.strip line
   invalid = InvalidPlaceholder "generated fragments must use a standalone \\tex2ssgenerated{name} line"
 
-validateFragmentName :: Text -> Aeson.Parser ()
-validateFragmentName name =
-  unless (validFragmentName name) $
-    fail "fragment names must match [a-z0-9][a-z0-9_-]*"
-
 validFragmentName :: Text -> Bool
 validFragmentName name =
   case Text.uncons name of
@@ -167,13 +248,6 @@ validFragmentName name =
   validFirst character =
     ('a' <= character && character <= 'z') || ('0' <= character && character <= '9')
   validRest character = validFirst character || character == '_' || character == '-'
-
-rejectUnknown :: String -> Set.Set Text -> Object -> Aeson.Parser ()
-rejectUnknown label allowed value =
-  let present = Set.fromList (map Key.toText $ KeyMap.keys value)
-      unknown = Set.toAscList (present `Set.difference` allowed)
-   in unless (null unknown) $
-        fail $ label <> " contains unknown fields: " <> Text.unpack (Text.intercalate ", " unknown)
 
 generatorDiagnostic :: Bundle -> FilePath -> PandocError -> Diagnostic
 generatorDiagnostic bundle scriptPath problem =